// infra/lib/enums.ts
export enum AppStage {
  SANDBOX = 'sandbox',
  STAGING = 'staging',
  PRODUCTION = 'prod'
} 

export enum AppStageProfiles {
  PRODUCTION = 'greenline-albert',
  STAGING = 'greenline-staging',
  SANDBOX = 'greenline-sandbox',
}

export enum DeploymentEnvironments {
  LOCAL = 'local',
  CI = 'CI'
}